Moved Game to Visuals
