+++ src/Games/Chess/visuals/Driver.java
@@ -1,8 +1,5 @@
 package Games.Chess.Visuals;
 
-
-import Games.Chess.Mechanics.Game;
-
 public class Driver {
 
     public static void main(String[] args) {
